organizing kudo code a bit more
