+++ src/commands/kudos.ts
@@ -4,8 +4,7 @@ import {
   CommandInteractionOptionResolver,
   GuildMember,
 } from 'discord.js'
-import { createKudo } from '../utils/db/kudos'
-import { getOrCreateUser } from '../utils/db/users'
+import { giveKudos } from '../utils/db/kudos'
 
 const choices = Object.keys(KudoCategory).map((key) => {
   return {
@@ -14,70 +13,46 @@ const choices = Object.keys(KudoCategory).map((key) => {
   }
 })
 
-const giveKudos = async (
+const handleKudos = async (
   interaction: CommandInteraction,
   options: CommandInteractionOptionResolver
 ) => {
-  await interaction.deferReply()
+  try {
 
-  const mentionedUser = options.getMentionable('user', false)
-  console.log(mentionedUser)
-  if (!(mentionedUser instanceof GuildMember)) {
+    await interaction.deferReply()
+
+    const mentionedUser = options.getMentionable('user', false)
+
+    if (!(mentionedUser instanceof GuildMember)) {
+      return interaction.editReply({
+        content: `You can only give kudos to a User 🥰`,
+      })
+    }
+
+    const { id: receiverId } = mentionedUser
+    const { id: giverId } = interaction.user
+
+
+    const description = options.getString('for', true);
+    const categoryString = options.getString('category', true)
+
+    //HACK: https://stackoverflow.com/questions/50417254/dynamically-access-enum-in-typescript-by-key
+    const category: KudoCategory =
+      KudoCategory[categoryString as keyof typeof KudoCategory]
+
+    giveKudos(giverId, receiverId, category, description);
     return interaction.editReply({
-      content: `You can only give kudos to a User 🥰`,
+      content: `${category} Kudos were given to ${mentionedUser.user.username} for '${description}'`,
+    })
+  } catch (err) {
+    console.error(err)
+    return interaction.editReply({
+      content: `Something went wrong giving kudos :(`,
     })
   }
-
-  const { id: receiverId } = mentionedUser
-  const receivingUser = await getOrCreateUser(receiverId)
-  console.log({ receivingUser })
-  //Just incase we aren't able to create the user lets leave the guard in place
-  if (!receivingUser) {
-    console.info("That user doesn't exist")
-    return interaction.editReply({
-      content: `Something wasn't right with those users`,
-    })
-  }
-
-  const { id: giverId } = interaction.user
-  const givingUser = await getOrCreateUser(giverId)
-
-  if (!givingUser) {
-    console.info("That user doesn't exist")
-    return interaction.editReply({
-      content: `Something wasn't right with those users`,
-    })
-  }
-
-  const description = options.getString('description')
-  const categoryString = options.getString('category', true)
-  //HACK: https://stackoverflow.com/questions/50417254/dynamically-access-enum-in-typescript-by-key
-  const category: KudoCategory =
-    KudoCategory[categoryString as keyof typeof KudoCategory]
-  console.log({ categoryString })
-
-  const kudo: Prisma.KudoCreateInput = {
-    category: category,
-    receiver: {
-      connect: {
-        id: receiverId,
-      },
-    },
-    giver: {
-      connect: {
-        id: giverId,
-      },
-    },
-  }
-  const createdKudo = await createKudo(kudo)
-  console.log(createdKudo)
-  console.log(receivingUser, givingUser, description)
-  return interaction.editReply({
-    content: `Kudos were given`,
-  })
 }
 export default {
-  callback: giveKudos,
+  callback: handleKudos,
   name: 'kudos',
   description: 'Give kudos to a community member',
   options: [
@@ -88,7 +63,7 @@ export default {
       type: 'MENTIONABLE',
     },
     {
-      name: 'description',
+      name: 'for',
       description: `What are you givin kudos for?`,
       required: true,
       type: 'STRING',
